requirements.txt
numpy
Pillow
pygame
opencv
